Stay in screen
changed the alignment behavior so that the agent will produce a vector given zero context

Created and added a stay in Radius Behavior

Added a function to set each agents volume to 1 / total number of agents. Thus I removed the gain plugin from the synths mixer and removed the boid 1 and 2 mixer busses.

Changed the freqAdjust method to select a random midi note frequency from an array in the midi controller